requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
apturl==0.5.2
asttokens==2.0.5
async-timeout==4.0.2
attrs==22.1.0
base58==2.1.1
bitarray==2.6.0
black==22.6.0
blinker==1.4
Brlapi==0.8.3
certifi==2022.6.15
chardet==4.0.0
charset-normalizer==2.1.0
chrome-gnome-shell==0.0.0
click==8.1.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cytoolz==0.12.0
dataclassy==0.11.1
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info===1.1build1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
eip712==0.1.0
eth-abi==2.2.0
eth-account==0.5.9
eth-brownie==1.19.1
eth-event==1.2.3
eth-hash==0.3.3
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
execnet==1.9.0
filelock==3.6.0
frozenlist==1.3.1
greenlet==1.1.2
hexbytes==0.2.2
httplib2==0.20.2
hypothesis==6.27.3
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.0
iniconfig==1.1.1
ipfshttpclient==0.8.0a2
jeepney==0.7.1
jsonschema==3.2.0
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-object-proxy==1.7.1
louis==3.20.0
lru-dict==1.1.8
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
more-itertools==8.10.0
msgpack==1.0.3
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==0.4.3
mythx-models==1.9.1
neovim==0.3.1
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==21.3
parsimonious==0.8.1
pathspec==0.9.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.2
pluggy==1.0.0
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
py==1.11.0
py-solc-ast==1.2.9
py-solc-x==1.1.1
pycairo==1.20.1
pycryptodome==3.15.0
pycryptodomex==3.11.0
pycups==2.0.1
Pygments==2.12.0
pygments-lexer-solidity==0.7.0
PyGObject==3.42.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.5.0
pynvim==0.4.2
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
pytest==6.2.5
pytest-forked==1.4.0
pytest-xdist==1.34.0
python-apt==2.3.0+ubuntu2.1
python-dateutil==2.8.1
python-debian===0.1.43ubuntu1
python-dotenv==0.16.0
pythx==1.6.1
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.28.1
rlp==2.0.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
semantic-version==2.8.5
six==1.16.0
sortedcontainers==2.4.0
systemd-python==234
texttable==1.6.4
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tqdm==4.64.0
ubuntu-advantage-tools==27.10
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.11
varint==1.0.2
virtualenv==20.13.0+ds
vvm==0.1.0
vyper==0.3.6
wadllib==1.3.6
wcwidth==0.2.5
web3==5.30.0
websocket-client==1.2.3
websockets==9.1
wrapt==1.14.1
xdg==5
xkit==0.0.0
yarl==1.8.1
youtube-dl==2021.12.17
zipp==1.0.0
